+++ .eslintrc.js
@@ -19,8 +19,10 @@ module.exports = {
         project: './tsconfig.json'
       },
       rules: {
+        '@typescript-eslint/block-spacing': 'error',
         '@typescript-eslint/consistent-type-exports': ['error'],
         '@typescript-eslint/consistent-type-imports': ['error'],
+        '@typescript-eslint/key-spacing': 'error',
         '@typescript-eslint/member-delimiter-style': [
           'error',
           {
@@ -35,8 +37,13 @@ module.exports = {
             multilineDetection: 'brackets'
           }
         ],
+        '@typescript-eslint/no-import-type-side-effects': 'error',
+        '@typescript-eslint/no-mixed-enums': 'error',
         '@typescript-eslint/no-unused-vars': 'error',
-        '@typescript-eslint/switch-exhaustiveness-check': 'error'
+        '@typescript-eslint/switch-exhaustiveness-check': 'error',
+
+        'block-spacing': 'off', // @typescript-eslint/block-spacing used instead
+        'key-spacing': 'off' // @typescript-eslint/key-spacing used instead
       }
     },
     {