+++ CHANGELOG.md
@@ -10,6 +10,13 @@ Versioning].
 
 - _No changes yet_
 
+## [1.0.0] - 2023-02-25
+
+- (`9ddf3fe`) _(Breaking)_ Rename the rule `no-top-level-side-effect` to
+  `no-top-level-side-effects`.
+- (`cec69d5`) _(Breaking)_ Report top-level IIFEs by default.
+- (`92a3f58`) Improve performance of `no-top-level-variables`.
+
 ## [0.3.0] - 2023-01-13
 
 - (`a85b514`) Add support for configuring if top-level Immediately Invoked